Solution to Input and Print a Double

+++ Input and Print a Double@@6788d5012db25efec946f942/code.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 
 int main() {
-    float f;
-    scanf("%f",&f);
-    printf("You entered: %.4lf", f);
+    double f;
+    scanf("%lf",&f);
+    printf("You entered: %.4f", f);
     return 0;
 }
